# New TS index node implementation.