--- a/go/src/koding/klient/machine/index/node/node_test.go
+++ b/go/src/koding/klient/machine/index/node/node_test.go
@@ -1,0 +1,12 @@
+package node_test
+
+import (
+	"testing"
+
+	"koding/klient/machine/index/node"
+)
+
+func TestNodeSort(t *testing.T) {
+	nodes := []*node.Node{}
+	_ = nodes
+}