This PR adds new node implementation to mount's index. The old one suffers from race problems causing mounts to crash with concurrent map access panics.

Performance changes:
```
benchmark                  old ns/op     new ns/op     delta
BenchmarkNodeLookup-4      252           305           +21.03%
BenchmarkNodeAdd-4         3300          2405          -27.12%
BenchmarkNodeDel-4         352           804           +128.41%
BenchmarkNodeForEach-4     7196          3666          -49.06%

benchmark                  old allocs     new allocs     delta
BenchmarkNodeLookup-4      2              1              -50.00%
BenchmarkNodeAdd-4         19             16             -15.79%
BenchmarkNodeDel-4         0              1              +Inf%
BenchmarkNodeForEach-4     35             36             +2.86%

benchmark                  old bytes     new bytes     delta
BenchmarkNodeLookup-4      64            48            -25.00%
BenchmarkNodeAdd-4         1397          608           -56.48%
BenchmarkNodeDel-4         0             32            +Inf%
BenchmarkNodeForEach-4     2983          1000          -66.48%
```

## Motivation and Context
Prevent mounts and klient from random crashes.

## How Has This Been Tested?
Unit tests. 

## Types of changes
- [x] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)